# Init etcd and apiserver per test case in scheduler_perf integration tests