--- a/test/integration/scheduler_perf/create.go
+++ b/test/integration/scheduler_perf/create.go
@@ -24,7 +24,6 @@ import (
 	"os"
 	"time"
 
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -141,18 +140,6 @@ func (c *createAny) create(tCtx ktesting.TContext, env map[string]any) {
 				return fmt.Errorf("namespace not set for %q, but %q has scope %q", c.TemplatePath, gk, mapping.Scope.Name())
 			}
 			_, err = resourceClient.Create(tCtx, obj, options)
-		}
-		if err == nil && shouldCleanup(tCtx) {
-			tCtx.CleanupCtx(func(tCtx ktesting.TContext) {
-				del := resourceClient.Delete
-				if mapping.Scope.Name() != meta.RESTScopeNameNamespace {
-					del = resourceClient.Namespace(c.Namespace).Delete
-				}
-				err := del(tCtx, obj.GetName(), metav1.DeleteOptions{})
-				if !apierrors.IsNotFound(err) {
-					tCtx.ExpectNoError(err, fmt.Sprintf("deleting %s.%s %s", obj.GetKind(), obj.GetAPIVersion(), klog.KObj(obj)))
-				}
-			})
 		}
 		return err
 	}