@macsko 2024-07-23
> But before we merge it, let's also simplify the unit test configurations to reduce the total runtime.

Do we have any strategy for that, e.g. what test cases we want to keep, how many of different test cases we want, what is our expected testing time?