@pohly 2024-07-23
I think we should use exactly one workload for each scenario when doing integration testing. The goal for the overall runtime is [5 minutes at max](https://github.com/kubernetes/kubernetes/issues/126202#issuecomment-2237100235), less would be better.

Personally, I think very simple workloads are sufficient for integration testing. We don't want "stress test" style running in a presubmit. Perhaps we can run those in the periodic job, using the `testing.Short` function to distinguish between the two, but I think that can wat.