@pohly 2024-07-24
I did some timing measurements on my desktop system.

With commit c15cdf74316bdcef1f79cffe429dcff024eb302b (= don't reuse etcd + apiserver):
```console
$ gotestsum -f testname -- -count=1 -run=TestScheduling ./test/integration/scheduler_perf/ | grep -v -e SKIP  -e 'disabled by label filter' -e '^ *$'
PASS test/integration/scheduler_perf.TestScheduling/SchedulingBasic/500Nodes (8.42s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingBasic (8.42s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPodAntiAffinity/500Nodes (8.35s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPodAntiAffinity (8.35s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingSecrets (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingInTreePVs (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingMigratedInTreePVs (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingCSIPVs/500Nodes (13.95s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingCSIPVs (13.95s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPodAffinity (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredPodAffinity/500Nodes (9.03s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredPodAffinity (9.03s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredPodAntiAffinity (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingDaemonset (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingNodeAffinity (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/TopologySpreading/500Nodes (11.29s)
PASS test/integration/scheduler_perf.TestScheduling/TopologySpreading (11.29s)
PASS test/integration/scheduler_perf.TestScheduling/PreferredTopologySpreading (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/MixedSchedulingBasePod/500Nodes (12.34s)
PASS test/integration/scheduler_perf.TestScheduling/MixedSchedulingBasePod (12.34s)
PASS test/integration/scheduler_perf.TestScheduling/PreemptionBasic (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/PreemptionPVs (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/Unschedulable (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithMixedChurn/1000Nodes (12.22s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithMixedChurn (12.22s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingRequiredPodAntiAffinityWithNSSelector (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredAntiAffinityWithNSSelector (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingRequiredPodAffinityWithNSSelector (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredAffinityWithNSSelector (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithNodeInclusionPolicy (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimTemplate/fast (9.23s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimTemplate (9.23s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithMultipleResourceClaims/fast (9.43s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithMultipleResourceClaims (9.43s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimTemplateStructured/fast (5.24s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimTemplateStructured (5.24s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimStructured/fast (6.49s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimStructured (6.49s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWhileGated (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingGatedPodsWithPodAffinityImpactForThroughput (0.00s)
PASS test/integration/scheduler_perf.TestScheduling (106.00s)
PASS test/integration/scheduler_perf
=== Skipped
DONE 90 tests, 49 skipped in 109.660s
```

Not all scenarios are covered because some didn't have small workloads defined.

Before (commit 1bb62cd2750 = reuse etc + apiserver):
```
PASS test/integration/scheduler_perf.TestScheduling/SchedulingBasic/500Nodes (17.79s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingBasic (17.79s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPodAntiAffinity/500Nodes (13.54s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPodAntiAffinity (13.54s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingSecrets (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingInTreePVs (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingMigratedInTreePVs (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingCSIPVs/500Nodes (25.30s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingCSIPVs (25.30s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPodAffinity (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredPodAffinity/500Nodes (18.76s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredPodAffinity (18.76s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredPodAntiAffinity (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingDaemonset (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingNodeAffinity (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/TopologySpreading/500Nodes (21.92s)
PASS test/integration/scheduler_perf.TestScheduling/TopologySpreading (21.92s)
PASS test/integration/scheduler_perf.TestScheduling/PreferredTopologySpreading (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/MixedSchedulingBasePod/500Nodes (22.77s)
PASS test/integration/scheduler_perf.TestScheduling/MixedSchedulingBasePod (22.77s)
PASS test/integration/scheduler_perf.TestScheduling/PreemptionBasic (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/PreemptionPVs (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/Unschedulable (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithMixedChurn/1000Nodes (19.71s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithMixedChurn (19.71s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingRequiredPodAntiAffinityWithNSSelector (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredAntiAffinityWithNSSelector (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingRequiredPodAffinityWithNSSelector (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredAffinityWithNSSelector (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWhileGated (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingGatedPodsWithPodAffinityImpactForThroughput (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimTemplate/fast (12.25s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimTemplate (12.25s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithMultipleResourceClaims/fast (12.25s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithMultipleResourceClaims (12.25s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimTemplateStructured/fast (8.21s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimTemplateStructured (8.21s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimStructured/fast (5.25s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimStructured (5.25s)
PASS test/integration/scheduler_perf.TestScheduling (186.68s)
PASS test/integration/scheduler_perf
=== Skipped
DONE 87 tests, 47 skipped in 191.389s
```

That's less tests because .... no idea.

When defining very small workloads and using those for integration testing (patch below), with commit 1bb62cd2750 = reuse etc + apiserver:
```
PASS test/integration/scheduler_perf.TestScheduling/SchedulingBasic/10Nodes (9.31s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingBasic (9.31s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPodAntiAffinity/10Nodes (9.19s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPodAntiAffinity (9.19s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingSecrets (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingInTreePVs (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingMigratedInTreePVs (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingCSIPVs/10Nodes (9.44s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingCSIPVs (9.44s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPodAffinity (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredPodAffinity/10Nodes (9.29s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredPodAffinity (9.29s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredPodAntiAffinity (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingDaemonset (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingNodeAffinity (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/TopologySpreading/10Nodes (9.38s)
PASS test/integration/scheduler_perf.TestScheduling/TopologySpreading (9.38s)
PASS test/integration/scheduler_perf.TestScheduling/PreferredTopologySpreading (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/MixedSchedulingBasePod/10Nodes (9.39s)
PASS test/integration/scheduler_perf.TestScheduling/MixedSchedulingBasePod (9.39s)
PASS test/integration/scheduler_perf.TestScheduling/PreemptionBasic (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/PreemptionPVs (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/Unschedulable (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithMixedChurn/10Nodes (9.33s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithMixedChurn (9.33s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingRequiredPodAntiAffinityWithNSSelector (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredAntiAffinityWithNSSelector (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingRequiredPodAffinityWithNSSelector (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredAffinityWithNSSelector (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWhileGated (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingGatedPodsWithPodAffinityImpactForThroughput (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimTemplate/fast (12.33s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimTemplate (12.34s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithMultipleResourceClaims/fast (12.19s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithMultipleResourceClaims (12.19s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimTemplateStructured/fast (8.19s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimTemplateStructured (8.19s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimStructured/fast (5.18s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimStructured (5.18s)
PASS test/integration/scheduler_perf.TestScheduling (111.63s)
PASS test/integration/scheduler_perf
=== Skipped
DONE 94 tests, 54 skipped in 115.317s
```

When combining both (smaller workloads, no reuse of etcd):
```
PASS test/integration/scheduler_perf.TestScheduling/SchedulingBasic/10Nodes (6.66s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingBasic (6.66s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPodAntiAffinity/10Nodes (6.37s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPodAntiAffinity (6.37s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingSecrets (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingInTreePVs (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingMigratedInTreePVs (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingCSIPVs/10Nodes (6.35s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingCSIPVs (6.35s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPodAffinity (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredPodAffinity/10Nodes (6.43s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredPodAffinity (6.43s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredPodAntiAffinity (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingDaemonset (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingNodeAffinity (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/TopologySpreading/10Nodes (6.69s)
PASS test/integration/scheduler_perf.TestScheduling/TopologySpreading (6.69s)
PASS test/integration/scheduler_perf.TestScheduling/PreferredTopologySpreading (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/MixedSchedulingBasePod/10Nodes (6.28s)
PASS test/integration/scheduler_perf.TestScheduling/MixedSchedulingBasePod (6.28s)
PASS test/integration/scheduler_perf.TestScheduling/PreemptionBasic (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/PreemptionPVs (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/Unschedulable (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithMixedChurn/10Nodes (7.85s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithMixedChurn (7.85s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingRequiredPodAntiAffinityWithNSSelector (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredAntiAffinityWithNSSelector (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingRequiredPodAffinityWithNSSelector (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingPreferredAffinityWithNSSelector (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithNodeInclusionPolicy (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimTemplate/fast (9.57s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimTemplate (9.57s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithMultipleResourceClaims/fast (9.38s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithMultipleResourceClaims (9.38s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimTemplateStructured/fast (5.32s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimTemplateStructured (5.32s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimStructured/fast (6.58s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWithResourceClaimStructured (6.58s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingWhileGated (0.00s)
PASS test/integration/scheduler_perf.TestScheduling/SchedulingGatedPodsWithPodAffinityImpactForThroughput (0.00s)
PASS test/integration/scheduler_perf.TestScheduling (77.48s)
PASS test/integration/scheduler_perf
=== Skipped
DONE 97 tests, 56 skipped in 81.275s
```

That's the fastest approach.

Here's the patch that I was using:
```patch
diff --git a/test/integration/scheduler_perf/config/performance-config.yaml b/test/integration/scheduler_perf/config/performance-config.yaml
index 9d5183d4666..da9959bd6cb 100644
--- a/test/integration/scheduler_perf/config/performance-config.yaml
+++ b/test/integration/scheduler_perf/config/performance-config.yaml
@@ -23,8 +23,14 @@
     countParam: $measurePods
     collectMetrics: true
   workloads:
-  - name: 500Nodes
+  - name: 10Nodes
     labels: [integration-test, fast]
+    params:
+      initNodes: 10
+      initPods: 10
+      measurePods: 20
+  - name: 500Nodes
+    labels: [fast]
     params:
       initNodes: 500
       initPods: 500
@@ -52,8 +58,14 @@
     collectMetrics: true
     namespace: sched-1
   workloads:
-  - name: 500Nodes
+  - name: 10Nodes
     labels: [integration-test, fast]
+    params:
+      initNodes: 10
+      initPods: 2
+      measurePods: 8
+  - name: 500Nodes
+    labels: [fast]
     params:
       initNodes: 500
       initPods: 100
@@ -174,8 +186,14 @@
     persistentVolumeClaimTemplatePath: config/templates/pvc.yaml
     collectMetrics: true
   workloads:
-  - name: 500Nodes
+  - name: 10Nodes
     labels: [integration-test, fast]
+    params:
+      initNodes: 10
+      initPods: 10
+      measurePods: 20
+  - name: 500Nodes
+    labels: [fast]
     params:
       initNodes: 500
       initPods: 500
@@ -236,8 +254,14 @@
     namespace: sched-1
     collectMetrics: true
   workloads:
-  - name: 500Nodes
+  - name: 10Nodes
     labels: [integration-test, fast]
+    params:
+      initNodes: 10
+      initPods: 10
+      measurePods: 20
+  - name: 500Nodes
+    labels: [fast]
     params:
       initNodes: 500
       initPods: 500
@@ -348,8 +372,14 @@
     podTemplatePath: config/templates/pod-with-topology-spreading.yaml
     collectMetrics: true
   workloads:
-  - name: 500Nodes
+  - name: 10Nodes
     labels: [integration-test, fast]
+    params:
+      initNodes: 10
+      initPods: 20
+      measurePods: 20
+  - name: 500Nodes
+    labels: [fast]
     params:
       initNodes: 500
       initPods: 1000
@@ -391,7 +421,6 @@
       measurePods: 2000
 
 - name: MixedSchedulingBasePod
-  labels: [performance]
   defaultPodTemplatePath: config/templates/pod-default.yaml
   workloadTemplate:
   - opcode: createNodes
@@ -426,13 +455,20 @@
     countParam: $measurePods
     collectMetrics: true
   workloads:
-  - name: 500Nodes
+  - name: 10Nodes
     labels: [integration-test, fast]
+    params:
+      initNodes: 10
+      initPods: 4
+      measurePods: 20
+  - name: 500Nodes
+    labels: [performance, fast]
     params:
       initNodes: 500
       initPods: 200
       measurePods: 1000
   - name: 5000Nodes
+    labels: [performance]
     params:
       initNodes: 5000
       initPods: 2000
@@ -544,8 +580,13 @@
     podTemplatePath: config/templates//pod-default.yaml
     collectMetrics: true
   workloads:
-  - name: 1000Nodes
+  - name: 10Nodes
     labels: [integration-test, fast]
+    params:
+      initNodes: 10
+      measurePods: 10
+  - name: 1000Nodes
+    labels: [fast]
     params:
       initNodes: 1000
       measurePods: 1000
```